+++ src/main.rs
@@ -1,6 +1,7 @@
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
 mod app;
+mod autopair;
 mod command_palette;
 mod comment;
 mod diff;